+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 3.10.0 NEXT
 
+* Introduce generic tag assertion to `Asn1Element`
+
+
 ### 3.9.0 (Supreme 0.4.0)
 
 * Move `Attestation` from Supreme to Indispensable