+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 3.10.0 NEXT
 
 * Introduce generic tag assertion to `Asn1Element`
+* Change CSR to take an actual `CryptoSignature` instead of a ByteArray
 
 
 ### 3.9.0 (Supreme 0.4.0)